feat: academic calendar controllers and routes (#81)
* feat: update academic_calendar, holiday and day_swap model

* feat: add controllers for academic_calendar, holiday and day_swap

* feat: add routes for academic_calendar, holiday and day_swap

* feat: add db seeder for academic_calendar, holiday and day_swap

* refactor: rename academic_calendar seeder name

* refactor: convert route URLs to snake_case

* fix: fix migrations types

* fix: fix models types and column_names

* feat: update seeder to libraries and other relations

* feat: add library controllers and update others

* feat: add routes for library controller

* fix: fix operator=

* refactor: simplify code

* refactor: move endpoints into prefix url

* feat: add typedModel decorator to library model

* fix: change type to enum in model's decorator

* fix: fix enum search in controllers and add relevant relations

+++ database/migrations/1732381079405_create_buildings_table.ts
@@ -17,7 +17,7 @@ export default class extends BaseSchema {
         })
         .notNullable();
 
-      table.bigInteger("campus_id").unsigned().notNullable();
+      table.integer("campus_id").unsigned().notNullable();
       table
         .foreign("campus_id")
         .references("id")
